renamed __init__.py to main.py to stop python from creating a 'src' package (should make various references simpler and avoid some import problems); updated everything downstream to work with that change
